Make the height map into a integer based one.

+++ src/fractal_heightmap.h
@@ -9,13 +9,13 @@
 struct d_heightmap {
 	int width;
 	int height;
-	float *array;
+	short *array;
 };
 
 void d_fractal_heightmap_generate (struct d_heightmap *hm, int seed, float heightScale, float h);
 struct d_heightmap* d_fractal_heightmap_new (int size);
 void d_fractal_heightmap_destroy (struct d_heightmap *hm);
 
-float d_fractal_heightmap_get (struct d_heightmap *hm, int x, int y);
+short d_fractal_heightmap_get (struct d_heightmap *hm, int x, int y);
 
 #endif
